pass dependents to getChangelogFunc
